Reach zero clang-tidy warnings, and fix two bugs it exposed
The remaining findings were mostly my own suppressions landing on the wrong
line after clang-format reflowed the code. Placed correctly, and along the way:

  - `add`'s by-value parameters were the copy clang-tidy meant, not `larger`'s
    const references; the NOLINT had been on the wrong function entirely
  - two more hand-rolled algorithm calls replaced with ranges versions
  - a redundant cast removed: bit_width already returns the right type
  - `width.name()` reached a static through an instance in 08.05's tests

Two real bugs surfaced by running the checks properly:

  - scripts/check-course.sh reused a build directory whatever options it had
    been configured with, so it reported failures that were its own. It now
    always reconfigures.
  - 02.05's starter does not compile, and nobody had said so. The reason is
    the lesson: a designated initialiser may only omit a member that has a
    default, so leaving out `.host` is
    -Wmissing-designated-field-initializers until the defaults exist. Now
    documented in the exercise's NOTE.

Final state: 0 clang-tidy warnings across the 100 solutions, clang-format
clean, 100/100 solutions passing under Debug, RelWithDebInfo, ASan+UBSan and
TSan, on both Apple Clang 21 and Homebrew Clang 22 -- and all 100 starters
still fail, as they must.

+++ solutions/12_modern_standard_library/05_bit_operations/exercise.cpp
@@ -18,7 +18,9 @@ constexpr std::uint32_t round_up_to_power_of_two(std::uint32_t value) {
 }
 
 constexpr int bits_needed(std::uint32_t value) {
-  return static_cast<int>(std::bit_width(value));
+  // bit_width returns the same width as its argument, so on a uint32_t the
+  // result is already an int -- no cast needed.
+  return std::bit_width(value);
 }
 
 // bit_cast checks at compile time that both types are trivially copyable and
